% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_promille.R
\name{get_alcohol}
\alias{get_alcohol}
\title{get the amount of pure alcohol someone drank}
\usage{
get_alcohol(drinks)
}
\arguments{
\item{drinks}{a named vector whose names are a subset of "massn", "hoibe",
"wein" and "schnaps" and whose entries contain the amount of consumed units}
}
\value{
the output is the amount of pure alcohol drank (in ml)
}
\description{
get the amount of pure alcohol someone drank
}
