% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_promille.R
\name{get_bodywater}
\alias{get_bodywater}
\title{get the body water and give error message in case of illegal drinking}
\usage{
get_bodywater(sex = c("male", "female"), age, height, weight, drinks)
}
\arguments{
\item{sex}{either male or female}

\item{age}{a number between 10 and 110}

\item{height}{a height between 100 and 230}

\item{weight}{a weight between 40 and 300}

\item{drinks}{the drinks that were consumed - the reason why this is to be
passed as an argument is only for a possible warning in case of illegal
drinking behvaiour}
}
\description{
get the body water and give error message in case of illegal drinking
}
