% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_promille.R
\name{get_permille}
\alias{get_permille}
\title{gets the blood alcohol concentratio in permille}
\usage{
get_permille(alcohol_drunk, bodywater, drinking_time)
}
\arguments{
\item{alcohol_drunk}{a number that indicates the amount of pure alcohol
drunk (in ml) - should be the output of get_alcohol}

\item{bodywater}{water in the body in liter
- should be the output of get_bodywater}

\item{drinking_time}{a sorted posixct vector of length 2 that contains the
beginning and the end of the drinking time}
}
\description{
a subfunction of tell_me_how_drunk that calculates
}
