% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_promille.R
\name{show_me_how_drunk}
\alias{show_me_how_drunk}
\title{plots the blood alcohol concentration in permille over time}
\usage{
show_me_how_drunk(age, sex, height, weight, drinking_time, drinks)
}
\arguments{
\item{age}{a number between 10 and 110}

\item{sex}{either "male" or "female"}

\item{height}{a number between 100 and 230}

\item{weight}{a number between 40 and 300}

\item{drinking_time}{a sorted posixct vector of length 2 that contains the
beginning and the end of the drinking time}

\item{drinks}{a named vector whose names are a subset of "massn", "hoibe",
"wein" and "schnaps" and whose entries contain the amount of consumed units}
}
\value{
the ggplot object
}
\description{
plots the blood alcohol concentration
}
